[core/out_http/out_stdout] De-duplicate msgpack to json
There were 2 identical functions msgpack_to_json()
in out_http and out_stdout. Refactor to flb_pack as a
library function.
