+++ include/fluent-bit/flb_time.h
@@ -22,13 +22,13 @@
 #define FLB_TIME_H
 
 #include <fluent-bit/flb_info.h>
-#include <fluent-bit/flb_time_utils.h>
 
 #include <time.h>
 #include <msgpack.h>
 struct flb_time {
     struct timespec tm;
 };
+#include <fluent-bit/flb_time_utils.h>
 
 /*
    to represent eventtime of fluentd
