+++ plugins/out_stdout/stdout.h
@@ -24,16 +24,11 @@
 #define FLB_STDOUT_OUT_MSGPACK      0
 #define FLB_STDOUT_OUT_JSON_LINES   1
 
-#define FLB_STDOUT_JSON_DATE_DOUBLE      0
-#define FLB_STDOUT_JSON_DATE_ISO8601     1
-#define FLB_STDOUT_JSON_DATE_ISO8601_FMT "%Y-%m-%dT%H:%M:%S"
-
 struct flb_out_stdout_config {
     int out_format;
 
-    int json_date_format;
+    flb_date_format_t json_date_format;
     char *json_date_key;
-    size_t json_date_key_len;
 };
 
 #endif
